remove legacy references
