+++ src/components/listitems/index.tsx
@@ -8,3 +8,4 @@ export * from "./ListItemTransaction";
 export * from "./PressableListItemsBase";
 export * from "./ListItemSwitch";
 export * from "./ListItemCheckbox";
+export * from "./ListItemRadio";
