+++ src/components/radio/RadioGroup.tsx
@@ -2,20 +2,20 @@ import React from "react";
 import { View } from "react-native";
 import { Divider } from "../divider";
 import { IOIcons } from "../icons";
-import { RadioListItem } from "./RadioListItem";
+import { ListItemRadio } from "../listitems/ListItemRadio";
 
 export type NewRadioItem<T> = {
-    id: T;
-    value: string;
-    description?: string;
-    icon?: IOIcons;
-    disabled?: boolean;
+  id: T;
+  value: string;
+  description?: string;
+  icon?: IOIcons;
+  disabled?: boolean;
 };
 
 type Props<T> = {
-    items: ReadonlyArray<NewRadioItem<T>>;
-    selectedItem?: T;
-    onPress: (selected: T) => void;
+  items: ReadonlyArray<NewRadioItem<T>>;
+  selectedItem?: T;
+  onPress: (selected: T) => void;
 };
 
 /**
@@ -24,20 +24,20 @@ type Props<T> = {
  * The item with the `id` equal to the `selectedItem` is the active one.
  */
 export const RadioGroup = <T,>({ items, selectedItem, onPress }: Props<T>) => (
-    <View>
-        {items.map((item, index) => (
-            <React.Fragment key={`radio_item_${item.id}`}>
-                <RadioListItem
-                    testID={`RadioItemTestID_${item.id}`}
-                    value={item.value}
-                    description={item.description}
-                    icon={item.icon}
-                    disabled={item.disabled}
-                    onValueChange={() => onPress(item.id)}
-                    selected={selectedItem === item.id}
-                />
-                {index < items.length - 1 && <Divider />}
-            </React.Fragment>
-        ))}
-    </View>
+  <View>
+    {items.map((item, index) => (
+      <React.Fragment key={`radio_item_${item.id}`}>
+        <ListItemRadio
+          testID={`RadioItemTestID_${item.id}`}
+          value={item.value}
+          description={item.description}
+          icon={item.icon}
+          disabled={item.disabled}
+          onValueChange={() => onPress(item.id)}
+          selected={selectedItem === item.id}
+        />
+        {index < items.length - 1 && <Divider />}
+      </React.Fragment>
+    ))}
+  </View>
 );
