+++ src/components/radio/index.tsx
@@ -1,3 +1,2 @@
 export * from "./AnimatedRadio";
 export * from "./RadioGroup";
-export * from "./RadioListItem";
